The complexity 56%
384 / 10000
АНГЛИЙСКИЙ
Перевести вGoogleBing
You are given a chain of N characters consisting of uppercase English letters. It is necessary to go from the first character of the chain to the last character, jumping no more than K characters. The cost of a jump in which the symbol does not change is 0, and the cost of jumping to another symbol is 1.

You want to write a program that calculates the lowest cost of the transition from the first to the last character.